refactor: Rename cogni_mcp.py to setup_connection_to_cogni_mcp.py
